New Windows, New Code. Getting Better

+++ src/main/java/ir/tic/clouddc/pm/PmDetailRepository.java
@@ -11,16 +11,16 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PmDetailRepository extends JpaRepository<PmDetail, Integer> {
-    Optional<PmDetail> findByPmIdAndActive(int pmId, boolean active);
+public interface PmDetailRepository extends JpaRepository<PmDetail, Long> {
+    Optional<PmDetail> findByPmIdAndActive(Long pmId, boolean active);
 
     @Query("SELECT p.pm FROM PmDetail p WHERE p.persistence.person.username = :username AND p.active = :active")
     List<Pm> fetchPmListByActivationAndPerson(@Nullable @Param("username") String username, @Param("active") boolean active);
 
     @Query("SELECT p.persistence.id FROM PmDetail p WHERE p.pm.id = :pmId")
-    List<Long> getPersistenceIdList(@Param("pmId") int pmId);
+    List<Long> getPersistenceIdList(@Param("pmId") Long pmId);
 
     @Query("SELECT p.persistence.person.username FROM PmDetail p WHERE p.pm.id = :pmId AND p.active = :active")
-    String fetchPmOwnerUsername(@Param("pmId") int pmId, @Param("active") boolean active);
+    String fetchPmOwnerUsername(@Param("pmId") Long pmId, @Param("active") boolean active);
 
 }
